requirements.txt
streamlit
pandas
supabase
python-dateutil
streamlit-aggrid
python-dotenv
